isso.utils.http: Use urllib instead of httplib
Rationale: When implementing OpenID, it's important to follow redirects,
and following redirects should be a good thing in general. So let's not
reinvent the wheel when urllib does that for us.

Small regression when using Python < 3.3: When checking the HTTP server
connection in make_app(), GET method will now be used instead of HEAD
(since explicit method selection was only added to urllib in Python 3.3),
wasting a small amount of bandwidth.