Add data-isso-reply-notifications to attributes
There is a header describing `data-isso-reply-notifications` but it is not in the initial data attributes examples. I assume the initial list is meant to be exhaustive? I overlooked this setting on my initial read through the docs because it wasn't there, so I thought it would be helpful to add.