Have client read out shared settings from server (#311)
The following settings are sent as a config object from the
server to the client:
- `reply-to-self`
- `require-author`
- `require-email`
- `reply-notifications`
- `gravatar`

This means that the following settings will be ignored when
set on the client side:
- `data-isso-reply-to-self`
- `data-isso-require-author`
- `data-isso-require-email`
- `data-isso-reply-notifications`
- `data-isso-gravatar`

Isso will notify the user on the browser console if a client
setting has been overwritten by a config from the server.

Gravatar setting trumps regular avatars:
In addition, `data-isso-avatar` will be set to `false` when
gravatars are enabled to prevent both regular avatars and
gravatars appearing side-by-side.

Documentation updates:
Update docs for shared server-client settings and clarify
documentation for gravatar/avatar settings