utils: Test remote_addr as bytes, not str
This was probably an oversight while porting to py2-only:

python2 support was removed in dd76aa0cf14e01dd7b2aff4342299b461f94f53e

Checking for not(str) and str at the same time is a no-op.
What makes sense is to check whether a byte object can be
decoded to unicode.