isso: config: Use isso/isso.cfg as default
`defaults.ini` was a symlink to `share/isso.cfg`. The
previous commits settled upon `isso.cfg` as a config
filename, so keep with that convention.

Move the config to `isso/` so that `pkg_resources` can find
it

This helps avoid confusion and cuts down the amount of
different config files flying around.