utils/html: Sanitize flags/exts earlier
Sanitization of config options should happen in `Markup`
class, not in `Sanitizer`. Extensions names are now also
sanitized in addition to flags.

Also filter out empty strings from `allowed-elements` and
`allowed-attributes` when the respective options in the
config file are empty.

Checking for `NoOptionError` was superfluous since the list
would always contain something when falling back to the
default config file, so removed here.