isso: css: Align action buttons vertically w/ inputs
The "Preview" and "Submit" buttons had
`display: inline-block` set, but this was ignored and reset
to `block` due to also being `float`-ed right.

The blocks would align at the same height as the labels, not
the height of the input fields. Because before the
introduction of the labels, the input fields and the submit
buttons aligned vertically, restore that state.

Set height of labels explicitly and offset the action
buttons by that same height vertically.

On mobile, where the post action buttons are on another row
below the input fields anyway, apply no top margin.