views: comments: Check using get(), not 'in'
Previously, if a dict item was supplied by the JS client but
contained an invalid value such as `null`, the server would
only check if the dict key was provided, not if its value
was falsy.

```python
>>> mydict = dict(a='hi', b=None)
>>> 'b' in mydict
True
>>> 'b' not in mydict
False
```

Instead, use `.get()` and check if the value is empty. Then,
for instance, attempt to parse thread uris if payload title
is invalid instead of simply accepting `null` as thread
title.

Fixes: https://github.com/posativ/isso/issues/874