migrate: Handle single newlines in WordPress comments
WordPress renders a single newline in a comment as a <br> tag, but Isso renders a
single newline in the comment as a single newline in the HTML. This is rendered
the same as if it was a space, all text on one line.

To fix, detect single newlines when importing WordPress comments and convert
to a line break in Markdown. Add a test for this also.

Example, this WordPress comment (as shown in CDATA of XML export):

> First line of comment.
> Second line of comment.

Renders in WordPress as:

> First line of comment.
> <br>
> Second line of comment.

But renders in Isso after import as if it was:

> First line of comment. Second line of comment.

After this commit is applied and comments re-imported, it renders as:

> First line of comment.
> Second line of comment.