Dockerfile: Explictly create volume paths and set permissions
This allows running the docker image with a non-root user (--user 2000)
with named/anonymous volumes (-v isso-db:/db).

When an empty volume is mounted into a container (non-bind mount), it is
prepopulated with the files at the mount path, including owners and
permissions. The issue is that since /db does not exist at docker build
time, the volume defaults to root ownership with 0755 permissions, which
prevents non-root users from creating the sqlite database. This change
explicitly creates the /config and /db directories with 1777 permissions
to allow any user to write to them.

Note that this has no impact on bind mounts, since these always use
permissions of the host filesystem.