docs: Document Isso Project Infrastructure
Try to codify the new reality of the project since Martin
stepped down and make sense of all the informal or
mostly-one-off processes we are currently running.

Brought about by recent events, I (@ix5) thought about the
nature of maintainership a bit and realized that we have a)
a pretty significant bus factor and b) our infra could be a
bit opaque to newcomers. Collecting all the information and
writing it down in one place also might help highlight a few
blind spots and areas for improvement.