Add ntfy as a notification backend
The SMTP notification backend looks quite complicated and is tricky to
setup so I wrote a ntfy backend based on what's in the Stdout
subscriber object. Everything has been tested to work on with my ntfy
setup, except for the new_thread callback, which i'm not sure exactly
how to trigger.

It is also currently missing documentations. I will add the relevant
sections to the documentation if the patch is considered useful.